Task 3: webpack config and modules added
