+++ public/index.html
@@ -10,7 +10,6 @@
     <link rel="stylesheet" href="css/normalize.css" />
     <link rel="stylesheet" href="css/style.css" />
     <title>TASKMANAGER</title>
-    <script defer src="../src/main.js"></script>
   </head>
 
   <body>
