+++ src/config/filters.js
@@ -1,0 +1,33 @@
+export default [
+  {
+    label: `ALL`,
+    count: 15,
+    checked: true
+  },
+  {
+    label: `OVERDUE`,
+    count: 0,
+    disabled: true,
+  },
+  {
+    label: `TODAY`,
+    count: 0,
+    disabled: true,
+  },
+  {
+    label: `FAVORITES`,
+    count: 7,
+  },
+  {
+    label: `Repeating`,
+    count: 2,
+  },
+  {
+    label: `Tags`,
+    count: 6,
+  },
+  {
+    label: `ARCHIVE`,
+    count: 115,
+  },
+];
