+++ src/config/index.js
@@ -1,0 +1,3 @@
+import filters from './filters';
+
+export {filters};
