+++ src/templates/index.js
@@ -1,0 +1,7 @@
+import filterTemplate from './filter';
+import cardTemplate from './card';
+
+export {
+  filterTemplate,
+  cardTemplate
+};
