+++ src/utils/index.js
@@ -1,0 +1,11 @@
+import {
+  getElement,
+  getAllElements,
+  renderElement
+} from './utils';
+
+export {
+  getElement,
+  getAllElements,
+  renderElement
+};
